Added more Pokemon for the first rendering on the home page, changes for IE 11 work perfectly
