+++ Daria Lobanova/pokedex/.gitignore
