+++ Daria Lobanova/pokedex/public/manifest.json
