+++ Daria Lobanova/pokedex/src/components/search/SearchBar.js
