--- Daria Lobanova/pokedex/src/components/variables.js
@@ -1,4 +1,0 @@
-export const URL = 'http://localhost:3333/pokemons';
-export const OPTIONS = { hour: 'numeric', minute: 'numeric' };
-export const PLACEHOLDER_SMALL = 'https://via.placeholder.com/250.png?text=Pokemon+Not+Found';
-export const PLACEHOLDER_LARGE = 'https://via.placeholder.com/475.png?text=Pokemon+Not+Found';
