+++ Daria Lobanova/pokedex/src/components/view/CatchedPage.js
