+++ Daria Lobanova/pokedex/src/components/view/Navbar.js
