+++ Daria Lobanova/pokedex/src/components/view/PokemonPage.js
