+++ Daria Lobanova/pokedex/src/store/constants.js
@@ -1,0 +1,2 @@
+export const ACTION_CHANGE_BUTTON = 'ACTION_CHANGE_BUTTON';
+export const ACTION_CHANGE_INPUT = 'ACTION_CHANGE_INPUT';
