--- Daria Lobanova/pokedex/src/store/variables.js
@@ -1,2 +1,0 @@
-export const ACTION_CHANGE_BUTTON = 'ACTION_CHANGE_BUTTON';
-export const ACTION_CHANGE_INPUT = 'ACTION_CHANGE_INPUT';
